refactor(contract): move confidential key derivation into api/ckd.rs
